ref: deleteBlock関数の作成
