+++ src/lib/utils/textareaUtils.ts
@@ -1,4 +1,5 @@
-import type { Block } from "@/types/type"
+import type { Block, TextAreaElementMap } from "@/types/type"
+import type { RefObject } from "react"
 
 export const adjustTextareaHeight = (textarea: HTMLTextAreaElement) => {
   textarea.style.height = "auto"
@@ -23,4 +24,17 @@ export const scrollElementIntoView = (elementId: string) => {
       console.error("指定された要素までスクロールできませんでした")
     }
   })
+}
+
+export const deleteBlock = (
+  refs: RefObject<TextAreaElementMap>,
+  blocks: Block[],
+  blockId: string,
+  focusTextarea: (blockId: string) => void
+) => {
+  const index = findIndexBlocks(blocks, blockId)
+  refs.current[blockId]?.remove()
+  blocks.splice(index, 1)
+  const prevBlockId = blocks[index - 1].id
+  focusTextarea(prevBlockId)
 }
